requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0
reportlab==4.0.7
PyPDF2==3.0.1
pdfkit==1.0.0
APScheduler==3.10.4
psycopg2-binary==2.9.9